packages/assistant-sdk/pyproject.toml
[project]
name = "assistant-sdk"
version = "0.1.0"
description = "Contracts and utilities for Assistant integrations"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "jinja2>=3.1.6",
    "python-frontmatter>=1.1.0",
    "jsonschema>=4.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant_sdk"]
